Module rustc_ast_passes::ast_validation

Structs

AstValidator 🔒

Enums

ForbiddenLetReason 🔒

Used to forbid let expressions in certain syntactic locations.

SelfSemantic 🔒

Is self allowed semantically as the first parameter in an FnDecl?

Constants

MORE_EXTERN 🔒

Functions

check_crate

deny_equality_constraints 🔒

When encountering an equality constraint in a where clause, emit an error. If the code seems like it’s setting an associated type, provide an appropriate suggestion.

validate_generic_param_order 🔒

Checks that generic parameters are in the correct order, which is lifetimes, then types and then consts. (<'a, T, const N: usize>)