Function rustc_ast_passes::ast_validation::deny_equality_constraints

fn deny_equality_constraints(
    this: &mut AstValidator<'_>,
    predicate: &WhereEqPredicate,
    generics: &Generics
)

When encountering an equality constraint in a where clause, emit an error. If the code seems like it’s setting an associated type, provide an appropriate suggestion.