Struct rustc_ast_passes::ast_validation::AstValidator

struct AstValidator<'a> {
    session: &'a Session,
    features: &'a Features,
    extern_mod: Option<&'a Item>,
    in_trait_impl: bool,
    in_const_trait_impl: bool,
    has_proc_macro_decls: bool,
    outer_impl_trait: Option<Span>,
    disallow_tilde_const: Option<DisallowTildeConstContext<'a>>,
    is_impl_trait_banned: bool,
    forbidden_let_reason: Option<ForbiddenLetReason>,
    lint_buffer: &'a mut LintBuffer,
}

Fields
session: &'a Sessionfeatures: &'a Featuresextern_mod: Option<&'a Item>

The span of the extern in an extern { ... } block, if any.

in_trait_impl: bool

Are we inside a trait impl?

in_const_trait_impl: boolhas_proc_macro_decls: boolouter_impl_trait: Option<Span>

Used to ban nested impl Trait, e.g., impl Into<impl Debug>. Nested impl Trait is allowed in associated type position, e.g., impl Iterator<Item = impl Debug>.

disallow_tilde_const: Option<DisallowTildeConstContext<'a>>is_impl_trait_banned: bool

Used to ban impl Trait in path projections like <impl Iterator>::Item or Foo::Bar<impl Trait>

forbidden_let_reason: Option<ForbiddenLetReason>

See ForbiddenLetReason

lint_buffer: &'a mut LintBuffer
Implementations

impl<'a> AstValidator<'a>

fn with_in_trait_impl( &mut self, is_in: bool, constness: Option<Const>, f: impl FnOnce(&mut Self) )

fn with_banned_impl_trait(&mut self, f: impl FnOnce(&mut Self))

fn with_tilde_const( &mut self, disallowed: Option<DisallowTildeConstContext<'a>>, f: impl FnOnce(&mut Self) )

fn with_tilde_const_allowed(&mut self, f: impl FnOnce(&mut Self))

fn with_banned_tilde_const( &mut self, ctx: DisallowTildeConstContext<'a>, f: impl FnOnce(&mut Self) )

fn with_let_management( &mut self, forbidden_let_reason: Option<ForbiddenLetReason>, f: impl FnOnce(&mut Self, Option<ForbiddenLetReason>) )

fn ban_let_expr(&self, expr: &'a Expr, forbidden_let_reason: ForbiddenLetReason)

Emits an error banning the let expression provided in the given location.

fn check_type_alias_where_clause_location( &mut self, ty_alias: &TyAlias ) -> Result<(), WhereClauseBeforeTypeAlias>

fn with_impl_trait(&mut self, outer: Option<Span>, f: impl FnOnce(&mut Self))

fn walk_ty(&mut self, t: &'a Ty)

fn visit_struct_field_def(&mut self, field: &'a FieldDef)

fn err_handler(&self) -> &Handler

fn check_lifetime(&self, ident: Ident)

fn check_label(&self, ident: Ident)

fn visibility_not_permitted( &self, vis: &Visibility, note: VisibilityNotPermittedNote )

fn check_decl_no_pat( decl: &FnDecl, report_err: impl FnMut(Span, Option<Ident>, bool) )

fn check_unnamed_field_ty(&self, ty: &Ty, span: Span)

fn deny_anon_struct_or_union(&self, ty: &Ty)

fn deny_unnamed_field(&self, field: &FieldDef)

fn check_trait_fn_not_const(&self, constness: Const)

fn check_fn_decl(&self, fn_decl: &FnDecl, self_semantic: SelfSemantic)

fn check_decl_num_args(&self, fn_decl: &FnDecl)

Emits fatal error if function declaration has more than u16::MAX arguments Error is fatal to prevent errors during typechecking

fn check_decl_cvaradic_pos(&self, fn_decl: &FnDecl)

fn check_decl_attrs(&self, fn_decl: &FnDecl)

fn check_decl_self_param(&self, fn_decl: &FnDecl, self_semantic: SelfSemantic)

fn check_defaultness(&self, span: Span, defaultness: Defaultness)

fn ending_semi_or_hi(&self, sp: Span) -> Span

If sp ends with a semicolon, returns it as a Span Otherwise, returns sp.shrink_to_hi()

fn check_type_no_bounds(&self, bounds: &[GenericBound], ctx: &str)

fn check_foreign_ty_genericless( &self, generics: &Generics, before_where_clause: &TyAliasWhereClause, after_where_clause: &TyAliasWhereClause )

fn check_foreign_kind_bodyless( &self, ident: Ident, kind: &str, body: Option<Span> )

fn check_foreign_fn_bodyless(&self, ident: Ident, body: Option<&Block>)

An fn in extern { ... } cannot have a body { ... }.

fn current_extern_span(&self) -> Span

fn check_foreign_fn_headerless( &self, ident: Ident, span: Span, header: FnHeader )

An fn in extern { ... } cannot have qualifiers, e.g. async fn.

fn check_foreign_item_ascii_only(&self, ident: Ident)

An item in extern { ... } cannot use non-ascii identifier.

fn check_c_variadic_type(&self, fk: FnKind<'a>)

Reject C-variadic type unless the function is foreign, or free and unsafe extern "C" semantically.

fn check_item_named(&self, ident: Ident, kind: &str)

fn check_nomangle_item_asciionly(&self, ident: Ident, item_span: Span)

fn check_mod_file_item_asciionly(&self, ident: Ident)

fn deny_generic_params(&self, generics: &Generics, ident: Span)

fn emit_e0568(&self, span: Span, ident: Span)

fn deny_super_traits(&self, bounds: &GenericBounds, ident_span: Span)

fn deny_where_clause(&self, where_clause: &WhereClause, ident_span: Span)

fn deny_items(&self, trait_items: &[P<AssocItem>], ident: Span)

fn correct_generic_order_suggestion(&self, data: &AngleBracketedArgs) -> String

fn check_generic_args_before_constraints(&self, data: &AngleBracketedArgs)

Enforce generic args coming before constraints in <...> of a path segment.

fn visit_ty_common(&mut self, ty: &'a Ty)

fn maybe_lint_missing_abi(&mut self, span: Span, id: NodeId)