rustc_ast_passes::ast_validation

Struct AstValidator

struct AstValidator<'a> {
    sess: &'a Session,
    features: &'a Features,
    extern_mod_span: Option<Span>,
    outer_trait_or_trait_impl: Option<TraitOrTraitImpl>,
    has_proc_macro_decls: bool,
    outer_impl_trait_span: Option<Span>,
    disallow_tilde_const: Option<TildeConstReason>,
    extern_mod_safety: Option<Safety>,
    extern_mod_abi: Option<ExternAbi>,
    lint_node_id: NodeId,
    is_sdylib_interface: bool,
    lint_buffer: &'a mut LintBuffer,
}

Fields
sess: &'a Sessionfeatures: &'a Featuresextern_mod_span: Option<Span>

The span of the extern in an extern { ... } block, if any.

outer_trait_or_trait_impl: Option<TraitOrTraitImpl>has_proc_macro_decls: boolouter_impl_trait_span: Option<Span>

Used to ban nested impl Trait, e.g., impl Into<impl Debug>. Nested impl Trait is allowed in associated type position, e.g., impl Iterator<Item = impl Debug>.

disallow_tilde_const: Option<TildeConstReason>extern_mod_safety: Option<Safety>

Used to ban explicit safety on foreign items when the extern block is not marked as unsafe.

extern_mod_abi: Option<ExternAbi>lint_node_id: NodeIdis_sdylib_interface: boollint_buffer: &'a mut LintBuffer
Implementations

impl<'a> AstValidator<'a>

fn with_in_trait_impl( &mut self, trait_: Option<(Const, ImplPolarity, &'a TraitRef)>, f: impl FnOnce(&mut Self), )

fn with_in_trait( &mut self, span: Span, constness: Const, f: impl FnOnce(&mut Self), )

fn with_in_extern_mod( &mut self, extern_mod_safety: Safety, abi: Option<ExternAbi>, f: impl FnOnce(&mut Self), )

fn with_tilde_const( &mut self, disallowed: Option<TildeConstReason>, f: impl FnOnce(&mut Self), )

fn check_type_alias_where_clause_location( &mut self, ty_alias: &TyAlias, ) -> Result<(), WhereClauseBeforeTypeAlias>

fn with_impl_trait( &mut self, outer_span: Option<Span>, f: impl FnOnce(&mut Self), )

fn walk_ty(&mut self, t: &'a Ty)

fn dcx(&self) -> DiagCtxtHandle<'a>

fn visibility_not_permitted( &self, vis: &Visibility, note: VisibilityNotPermittedNote, )

fn check_decl_no_pat( decl: &FnDecl, report_err: impl FnMut(Span, Option<Ident>, bool), )

fn check_trait_fn_not_const(&self, constness: Const, parent: &TraitOrTraitImpl)

fn check_async_fn_in_const_trait_or_impl( &self, sig: &FnSig, parent: &TraitOrTraitImpl, )

fn check_fn_decl(&self, fn_decl: &FnDecl, self_semantic: SelfSemantic)

fn check_decl_num_args(&self, fn_decl: &FnDecl)

Emits fatal error if function declaration has more than u16::MAX arguments Error is fatal to prevent errors during typechecking

fn check_decl_cvariadic_pos(&self, fn_decl: &FnDecl)

Emits an error if a function declaration has a variadic parameter in the beginning or middle of parameter list. Example: fn foo(..., x: i32) will emit an error.

fn check_decl_attrs(&self, fn_decl: &FnDecl)

fn check_decl_self_param(&self, fn_decl: &FnDecl, self_semantic: SelfSemantic)

fn check_extern_fn_signature( &self, abi: ExternAbi, ctxt: FnCtxt, ident: &Ident, sig: &FnSig, )

Check that the signature of this function does not violate the constraints of its ABI.

fn reject_safe_fn(&self, abi: ExternAbi, ctxt: FnCtxt, sig: &FnSig)

fn reject_coroutine(&self, abi: ExternAbi, sig: &FnSig)

fn reject_params_or_return(&self, abi: ExternAbi, ident: &Ident, sig: &FnSig)

fn check_item_safety(&self, span: Span, safety: Safety)

This ensures that items can only be unsafe (or unmarked) outside of extern blocks.

This additionally ensures that within extern blocks, items can only be safe/unsafe inside of a unsafe-adorned extern block.

fn check_fn_ptr_safety(&self, span: Span, safety: Safety)

fn check_defaultness(&self, span: Span, defaultness: Defaultness)

fn ending_semi_or_hi(&self, sp: Span) -> Span

If sp ends with a semicolon, returns it as a Span Otherwise, returns sp.shrink_to_hi()

fn check_type_no_bounds(&self, bounds: &[GenericBound], ctx: &str)

fn check_foreign_ty_genericless( &self, generics: &Generics, where_clauses: &TyAliasWhereClauses, )

fn check_foreign_kind_bodyless( &self, ident: Ident, kind: &str, body_span: Option<Span>, )

fn check_foreign_fn_bodyless(&self, ident: Ident, body: Option<&Block>)

An fn in extern { ... } cannot have a body { ... }.

fn current_extern_span(&self) -> Span

fn check_foreign_fn_headerless(&self, _: FnHeader)

An fn in extern { ... } cannot have qualifiers, e.g. async fn.

fn check_foreign_item_ascii_only(&self, ident: Ident)

An item in extern { ... } cannot use non-ascii identifier.

fn check_c_variadic_type(&self, fk: FnKind<'a>)

Reject invalid C-variadic types.

C-variadics must be:

Non-const

Either foreign, or free and unsafe extern "C" semantically

fn check_item_named(&self, ident: Ident, kind: &str)

fn check_nomangle_item_asciionly(&self, ident: Ident, item_span: Span)

fn check_mod_file_item_asciionly(&self, ident: Ident)

fn deny_generic_params(&self, generics: &Generics, ident_span: Span)

fn deny_super_traits(&self, bounds: &GenericBounds, ident: Span)

fn deny_where_clause(&self, where_clause: &WhereClause, ident: Span)

fn deny_items(&self, trait_items: &[Box<AssocItem>], ident_span: Span)

fn correct_generic_order_suggestion(&self, data: &AngleBracketedArgs) -> String

fn check_generic_args_before_constraints(&self, data: &AngleBracketedArgs)

Enforce generic args coming before constraints in <...> of a path segment.

fn visit_ty_common(&mut self, ty: &'a Ty)

fn handle_missing_abi(&mut self, span: Span, id: NodeId)

fn visit_attrs_vis(&mut self, attrs: &'a AttrVec, vis: &'a Visibility)

fn visit_attrs_vis_ident( &mut self, attrs: &'a AttrVec, vis: &'a Visibility, ident: &'a Ident, )